tests: fix spurious failure on Mac OS X
See automake bug#11306.

* t/lex-libobj.sh: We used to pass LEXLIB='-L /lib' to the ./configure
invocation, to prevent it from explicitly searching for a "lex library".
But the linker on MacOS X 10.5 apparently chokes on that:

  cc -g -O2  -o foo foo.o -L /lib yywrap.o
  ld: file not found: /lib
  collect2: ld returned 1 exit status

So, we now instead pass LEXLIB=' ', which being non-empty is enough
to prevent the search of a "lex library".  And since we are at it,
enhance comments on this non-obvious usage.

Co-authored-by: Bruno Haible <bruno@clisp.org>
Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>