[ng] dist: simplify VPATH handling
Now that we don't need to cater to Sun and Tru64 make, we can simplify
and rationalize our VPATH support in the creation of the distribution
directory.  Since we are at it, we clarify some existing comments, add
some new ones, and remove the obsolete ones.

This change has a little collateral effect in that after it the use of
"bare" wildcards in EXTRA_DIST does not work as intended anymore:

    # Won't distribute all test files in the builddir anymore.
    EXTRA_DIST = *.test

One will have to use the $(wildcard) GNU make builtin instead, as in:

    EXTRA_DIST = $(wildcard *.test)

This new limitation is deemed acceptable, especially because "bare"
wildcards suffer of a number of issues and inconsistencies, and their
use is mostly deprecated in favor of the $(wildcard) builtin (refer
to the GNU make manual for more details).

* lib/am/distdir.am (distdir): Extend the comments explaining why, while
filling up the dist directory, we need to check whether any distributed
file already exists in there; in particular, refer to some relevant bug
numbers and past commit IDs.  Remove workarounds and hack required to
support automatic VPATH rewrites with Sun and Tru64 make.  Refactor the
recipe using ...
(am__dist_files_1, am__dist_files_2, am__dist_files,
am__dist_parent_dirs): This new internal custom GNU make macros.
* NG-NEWS: Document that "bare" wildcards are not supported anymore
in EXTRA_DIST definition.
* t/extra12.sh: Adjust accordingly.
* t/nodep.sh: Remove, it's giving too much false positives to be
usable.
* t/dist-srcdir.sh: New test.
* t/dist-srcdir2.sh: New test, xfailing.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>