[ng] VarDef: store comments and values as a perl array
This is a preparatory refactoring in view of a planned change to
how the definitions of make variables augmented with '+=' are
output in the generated Makefiles.

* t/comments-in-var-def.sh: New xfailing test, shows an example of the
improved semantic for variable definitions/additions we want to reach
eventually.
* Makefile.am (XFAIL_TESTS): Add it.
* lib/Automake/VarDef.pm (new): Store values and comments for the
variable in array references, not in scalars.
(append): Just append the passed value and comment to those arrays,
without preprocessing.  The existing preprocessing has been moved ...
(raw_value): ... to this accessor function.
(comment, dump): Adjust.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>