dist: don't bother putting README first in $(DIST_COMMON)
Comments on some of our automake-time pre-processing of $(DIST_COMMON)
said that it was done in order to "put README first because it then
becomes easier to make a Usenet-compliant shar file".  But such a
format is hardly relevant anymore, and not worth the (albeit small)
added complexity.

* automake.in (handle_dist): Don't sort @dist_common.
(for_dist_common): Delete this function, is not used anymore.
* lib/am/distdir.am (DISTFILES): Remove obsolete comment.
* t/distcom4.sh: Adjust.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>