m4: prepend m4 builtins with "m4_"
The comments in 'lib/autoconf/autoconf.m4' from Autoconf 2.69 read:

  We discourage the use of the non prefixed macro names: M4sugar maps
  all the builtins into 'm4_'.  Autoconf has been converted to these
  names too.  But users may still depend upon these, so reestablish
  them.

Alas, in our init.m4 files, we were still using m4 builtins without
prefixing them with "m4_".  Not nice for a package that is one of the
major clients of Autoconf, and an integral part of the GNU Autotools!

Fix the issue.

* m4/init.m4 (AM_INIT_AUTOMAKE): Use 'm4_define' and 'm4_defn' instead
of 'define' and 'defn'.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>