self tests: cater for /bin/ksh symlinked to Zsh
* t/self-check-reexec.tap: When searching for a suitable non-Bash
shells, be sure to reject any shell that is Zsh "in disguise" (as
can be found on some Debian systems, where /bin/ksh can be symlinked
to /bin/zsh4).  This is required because our testsuite does not
support older versions of Zsh, and that was causing the test to fail
in the setup just described.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>