[ng] doc: reference "Automake-NG" rather than "Automake" in few places
We only do the change in few key places, which are the most relevant
and visible ones, and which ar also very unlikely to change.  We do
not do the Automake -> Automake-NG substitution throughout, because
we don't want to cause extra conflicts or inconsistency when merging
changes from the 'master' branch into 'ng/master'.

* doc/automake.texi: Reference "Automake-NG" rather than "Automake" in
few key places, to minimize the possibility of confusing the user.
In particular, do so in the '@top', '@setfilename' and '@settitle'
declarations, in the "front-cover", and in the introduction.
(PACKAGE_BUGREPORT): Redefine to 'automake-ng@gnu.org' rather than
to 'bug-automake@gnu.org'.
(Reporting Bugs): A bug reported to 'automake-ng@gnu.org' will not
cause anymore an issue to be opened automatically in the Automake
bug tracker; account for this.
Remove any reference to the old Automake Gnats database.
Adjust the list of authors, adding Ralf Wildenhues and myself (Stefano
Lattarini).
* doc/amhello/README: Adjust to match what is reported in the manual.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>