[ng] acconfig.h: don't support anymore (distribution and dependencies)
Support for acconfig.h is still present in Autoconf (version 2.69 at the
time of writing), but deprecated by a long time in favour of AH_VERBATIM
and AH_TEMPLATE.  The deprecation started at least in September 1999 (see
Changelog entry "Start to get rid of acconfig.h.  It is an anachronism",
dated 1999-09-28), and was basically completed by October 1999 (see the
ChangeLog entry "Moving most of the task of creating config.h.in from sh
to m4", dated 1999-10-01), i.e., before the Autoconf 2.50 release (dated
2001-05-21).

For more information, see chapter "Obsolete Constructs", section
"acconfig.h" of the Autoconf manual.  See also the discussion on automake
bug#7919, in particular the message <http://debbugs.gnu.org/7819#20>.

* automake.in (handle_configure): Don't automatically distribute the
'acconfig.h' file if it exists, and don't add it to the transform
'%FILES%' when processing the 'remake-hdr.am' Makefile fragment.
(@common_sometimes): Don't list 'acconfig.h' anymore.  Reformat it
while we are at it, to make it easier to add or remove entries in
the future.
* t/autodist-acconfig-no-subdir.sh: Remove as obsolete.
* t/autodist-acconfig: Likewise.
* t/confh.sh: Adjust.
* lib/missing (case $program in autoheader*)): Remove any mention
of 'acconfig.h'.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>